+++ driver/pcie/configration-space/info.sh
@@ -7,4 +7,7 @@ for config in ${configs[@]}
 do
 	echo "${config}"
 	hexdump -C ${config}
+	if [[ -e configuration-space ]]; then
+		./configuration-space --config-file ${config}
+	fi
 done